05Size/requirements.txt
mysql-connector-python
logging
argparse
black
opencv-contrib-python
#dautil
#matplotlib